add support for injected model classes
Allowing an easy injection mechanism for the java model representation
of the objects in the FreeIPA instance's LDAP database seems to make
the most sense, given the following two facts:

1. The LDAP schema backing FreeIPA instances is intentionally
customizable, making custom model object implementations a very likely
requirement for consumers of the Iris library.
2. The Tolerant Reader pattern seems well suited to the relationship
between this library and FreeIPA, so migrating the responsibility of
the FreeIPA model representation to the consuming application allows
the consumer to minimize modeling of unnecessary FreeIPA object
attributes.